requirements.txt
opencv-python>=4.8
numpy>=1.24
mss>=9.0.1
pillow>=10.0
pyperclip>=1.8.2
pynput>=1.7.6
pytesseract>=0.3.10

